Refactoring and moving files around
