+++ src/NServiceBus.Transport.SQS/Envelopes/Translators/IMessageEnvelopeTranslator.cs
@@ -2,7 +2,7 @@ namespace NServiceBus.Transport.SQS.Envelopes;
 
 using Amazon.SQS.Model;
 
-interface IMessageEnvelopeTranslator
+interface IMessageTranslator
 {
     TranslatedMessage TryTranslateIncoming(Message message, string messageIdOverride);
     OutgoingMessageTranslationResult TryTranslateOutgoing(OutgoingMessage message);
